+++ lib/drivers/battery/tps65987.cpp
@@ -40,84 +40,84 @@ template class TPS65987::Controller<I2C::Controller>;
  *
  * @return The raw PDO in bits 0-31.
  */
-unsigned long TPS65987::PDO::get_data(void) const
+uint32_t TPS65987::PDO::get_data() const
 {
     return this->data;
-};
+}
 /**
  * @brief Returns the maximum current indicated by the PDO.
  *
  * @return The maximum current in [mA].
  */
-unsigned int TPS65987::PDO::current(void) const
+uint16_t TPS65987::PDO::current() const
 {
     return (this->data & 0x3FF) * 10;
-};
+}
 
 /**
  * @brief Returns the fixed voltage indicated by the PDO.
  *
  * @return The voltage in [mV].
  */
-unsigned int TPS65987::PDO::voltage(void) const
+uint16_t TPS65987::PDO::voltage() const
 {
     // Get the bit position of the voltage depending on the PDO type
-    unsigned char bit_pos = this->type() == Fixed_Supply ? 10 : 20;
+    const uint8_t bit_pos = this->type() == Fixed_Supply ? 10 : 20;
 
     // Return the converted voltage in [mV]
     return ((this->data >> bit_pos) & 0x3FF) * 50;
-};
+}
 
 /**
  * @brief Check the type of PDO object.
  *
  * @return Returns the type of the object.
  */
-TPS65987::PDO::Type TPS65987::PDO::type(void) const
+TPS65987::PDO::Type TPS65987::PDO::type() const
 {
     return static_cast<Type>((this->data >> 30) & 0b11);
-};
+}
 
 /**
  * @brief Set the voltage of the PDO.
  *
  * @param voltage The voltage in [mV].
  */
-void TPS65987::PDO::set_voltage(const unsigned int voltage)
+void TPS65987::PDO::set_voltage(const uint16_t voltage)
 {
     // Calculate the voltage in 50mV steps
-    unsigned int volt = voltage / 50;
+    const uint16_t volt = voltage / 50;
 
     // Clear the voltage bits
     this->data &= ~(0x3FF << 10);
 
     // Set the voltage bits
     this->data |= (volt & 0x3FF) << 10;
-};
+}
 
 /**
  * @brief Set the current of the PDO.
  *
  * @param current The current in [mA].
  */
-void TPS65987::PDO::set_current(const unsigned int current)
+void TPS65987::PDO::set_current(const uint16_t current)
 {
     // Calculate the current in 10mA steps
-    unsigned int I = current / 10;
+    const uint16_t I_set = current / 10;
 
     // Clear the current bits
     this->data &= ~0x3FF;
 
     // Set the voltage bits
-    this->data |= (I & 0x3FF);
-};
+    this->data |= (I_set & 0x3FF);
+}
 
 /**
  * @brief Initialize the PD IC based on its current mode.
  * @return Returns True when the IC was initialized successfully.
  */
 template <class bus_controller>
-bool TPS65987::Controller<bus_controller>::initialize(void)
+bool TPS65987::Controller<bus_controller>::initialize()
 {
     // set the i2c address
     Bus::change_address(this->mybus, TPS65987::i2c_address);
@@ -136,7 +136,7 @@ bool TPS65987::Controller<bus_controller>::initialize(void)
     }
 
     return true;
-};
+}
 
 /**
  * @brief Read a register of the PD controller.
@@ -153,7 +153,7 @@ bool TPS65987::Controller<bus_controller>::read_register(const reg_t reg)
     // Send the data buffer to the register
     this->buffer_data[0] = reg.address;
     return Bus::read_array(this->mybus, reg.address, &this->buffer_data[1], reg.length + 1);
-};
+}
 
 /**
  * @brief Write a register of the PD controller.
@@ -171,7 +171,7 @@ bool TPS65987::Controller<bus_controller>::write_register(const reg_t reg)
     this->buffer_data[0] = reg.address;
     this->buffer_data[1] = reg.length;
     return Bus::send_array(this->mybus, &this->buffer_data[0], reg.length + 2);
-};
+}
 
 /**
  * @brief Read the currently active command and update the
@@ -179,7 +179,7 @@ bool TPS65987::Controller<bus_controller>::write_register(const reg_t reg)
  * @return Returns True when the command register was read successfully.
  */
 template <class bus_controller>
-bool TPS65987::Controller<bus_controller>::read_active_command(void)
+bool TPS65987::Controller<bus_controller>::read_active_command()
 {
     // Read the data
     if (!this->read_register(Register::Cmd1))
@@ -191,7 +191,7 @@ bool TPS65987::Controller<bus_controller>::read_active_command(void)
     this->cmd_active[1] = this->buffer_data[3];
     this->cmd_active[0] = this->buffer_data[4];
     return true;
-};
+}
 
 /**
  * @brief Write a command to the command register
@@ -206,19 +206,20 @@ bool TPS65987::Controller<bus_controller>::write_command(const char *cmd)
     // Assemble the cmd array
     this->buffer_cmd[0] = Register::Cmd1.address;
     this->buffer_cmd[1] = Register::Cmd1.length;
-    for (unsigned char nByte = 0; nByte < 4; nByte++)
-        this->buffer_cmd[nByte + 2] = *(cmd + nByte);
+    for (uint8_t nByte = 0; nByte < 4; nByte++)
+        this->buffer_cmd.at(nByte + 2) = *(cmd + nByte); // NOLINT(cppcoreguidelines-pro-bounds-pointer-arithmetic)
 
     // Send the array
-    unsigned char *dest = reinterpret_cast<unsigned char *>(&this->buffer_cmd[0]);
+    auto *dest = reinterpret_cast<std::uint8_t *>(&this->buffer_cmd[0]); // NOLINT
     return Bus::send_array(this->mybus, dest, Register::Cmd1.length + 2);
-};
+}
+
 /**
  * @brief Read the active mode of the controller.
  * @return Returns True when the mode was read successfully.
  */
 template <class bus_controller>
-bool TPS65987::Controller<bus_controller>::read_mode(void)
+bool TPS65987::Controller<bus_controller>::read_mode()
 {
     // Read the mode from the controller
     if (!this->read_register(Register::Mode))
@@ -239,7 +240,7 @@ bool TPS65987::Controller<bus_controller>::read_mode(void)
     default:
         this->mode_active = Mode::OTHER;
         break;
-    };
+    }
     return true;
 }
 
@@ -248,7 +249,7 @@ bool TPS65987::Controller<bus_controller>::read_mode(void)
  * @return Returns True when the status was read successfully.
  */
 template <class bus_controller>
-bool TPS65987::Controller<bus_controller>::read_PD_status(void)
+bool TPS65987::Controller<bus_controller>::read_PD_status()
 {
     // Read the status register
     if (!this->read_register(Register::PD_Status))
@@ -261,14 +262,14 @@ bool TPS65987::Controller<bus_controller>::read_PD_status(void)
     this->i2c_data.byte[0] = this->buffer_data[4];
 
     // Check plug type
-    unsigned char temp = this->i2c_data.byte[0];
-    if (temp & PlugDetails_0)
+    std::uint8_t temp = this->i2c_data.byte[0];
+    if ((temp & PlugDetails_0) != 0)
         this->contract_active.USB_type = 2;
     else
         this->contract_active.USB_type = 3;
 
     // Active role
-    if (temp & PresentRole)
+    if ((temp & PresentRole) != 0)
         this->contract_active.role = 1;
     else
         this->contract_active.role = 0;
@@ -300,30 +301,30 @@ bool TPS65987::Controller<bus_controller>::read_PD_status(void)
         this->contract_active.voltage = 0;
         this->contract_active.current = 0;
         break;
-    };
+    }
     return true;
-};
+}
 
 /**
  * @brief Read the status register [0x1A] from the PD Controller.
  * @return Returns True and the value of the lower 4 bytes of the status when read successfully.
  */
 template <class bus_controller>
-std::optional<unsigned long> TPS65987::Controller<bus_controller>::read_status(void)
+std::optional<uint32_t> TPS65987::Controller<bus_controller>::read_status()
 {
     // Read the status register
     if (!this->read_register(Register::Status))
         return {};
 
     // Convert the received data
-    unsigned long status = this->buffer_data[8];
+    uint32_t status = this->buffer_data[8];
     status |= (this->buffer_data[7] << 8);
     status |= (this->buffer_data[6] << 16);
     status |= (this->buffer_data[5] << 24);
 
     // Return the status
     return status;
-};
+}
 
 /**
  * @brief Read the active PDO from the PD Controller.
@@ -332,33 +333,32 @@ std::optional<unsigned long> TPS65987::Controller<bus_controller>::read_status(v
  * @return True and the active PDO when read successfully.
  */
 template <class bus_controller>
-std::optional<TPS65987::PDO> TPS65987::Controller<bus_controller>::read_active_pdo(void)
+std::optional<TPS65987::PDO> TPS65987::Controller<bus_controller>::read_active_pdo()
 {
     // Read the active PDO register
     if (!this->read_register(Register::Active_PDO))
         return {};
 
     // Convert the received data
-    unsigned long pdo = this->buffer_data[6];
+    uint32_t pdo = this->buffer_data[6];
     pdo |= (this->buffer_data[5] << 8);
     pdo |= (this->buffer_data[4] << 16);
     pdo |= (this->buffer_data[3] << 24);
 
     // Return the PDO
     return PDO(pdo);
-};
+}
 
 /**
  * @brief Read a TX source PDO from the PD Controller.
- * 
+ *
  * @tparam bus_controller The used bus for the communication.
  * @param pdo_number The PDO number to read.
  * @return Returns True and the PDO when read successfully and when the PDO number is valid.
  */
 template <class bus_controller>
 std::optional<TPS65987::PDO> TPS65987::Controller<bus_controller>::read_TX_sink_pdo(
-    const unsigned char pdo_number
-)
+    const std::uint8_t pdo_number)
 {
     // Read the TX sink PDO capabilities register
     if (!this->read_register(Register::TX_Sink_Cap))
@@ -370,10 +370,10 @@ std::optional<TPS65987::PDO> TPS65987::Controller<bus_controller>::read_TX_sink_
         return {};
 
     // Convert the received data
-    unsigned long pdo = this->buffer_data[56 - pdo_number * 4];
-    pdo |= (this->buffer_data[55 - pdo_number * 4] << 8);
-    pdo |= (this->buffer_data[54 - pdo_number * 4] << 16);
-    pdo |= (this->buffer_data[53 - pdo_number * 4] << 24);
+    uint32_t pdo = this->buffer_data.at(56 - pdo_number * 4);
+    pdo |= (this->buffer_data.at(55 - pdo_number * 4) << 8);
+    pdo |= (this->buffer_data.at(54 - pdo_number * 4) << 16);
+    pdo |= (this->buffer_data.at(53 - pdo_number * 4) << 24);
 
     return PDO(pdo);
-};
+}