+++ lib/drivers/battery/tps65987.h
@@ -22,30 +22,32 @@
 #define TPS65987_H_
 
 // === includes ===
-#include "drivers.h"
+// #include <stdint.h>
 #include <array>
+#include "drivers.h"
+#include "misc/bits.h"
 
 // === Command codes ===
 namespace TPS65987
 {
 
     // === Constants ===
-    constexpr unsigned char i2c_address = 0x40;
+    constexpr uint8_t i2c_address = 0x40;
 
     // === Bits ===
-    constexpr unsigned char PlugDetails_0 = (1 << 0);
-    constexpr unsigned char PlugDetails_1 = (1 << 1);
-    constexpr unsigned char CCPullUp_0 = (1 << 2);
-    constexpr unsigned char CCPullUp_1 = (1 << 3);
-    constexpr unsigned char PortType_0 = (1 << 4);
-    constexpr unsigned char PortType_1 = (1 << 5);
-    constexpr unsigned char PresentRole = (1 << 6);
+    constexpr uint8_t PlugDetails_0 = (1 << 0);
+    constexpr uint8_t PlugDetails_1 = (1 << 1);
+    constexpr uint8_t CCPullUp_0 = (1 << 2);
+    constexpr uint8_t CCPullUp_1 = (1 << 3);
+    constexpr uint8_t PortType_0 = (1 << 4);
+    constexpr uint8_t PortType_1 = (1 << 5);
+    constexpr uint8_t PresentRole = (1 << 6);
 
     // === Registers ===
     struct reg_t
     {
-        unsigned char address;
-        unsigned char length;
+        uint8_t address;
+        uint8_t length;
     };
 
     namespace Register
@@ -66,7 +68,7 @@ namespace TPS65987
     }; // namespace Register
 
     // === Modes ===
-    enum class Mode : unsigned char
+    enum class Mode : uint8_t
     {
         BOOT = 0,
         PTCH,
@@ -77,10 +79,10 @@ namespace TPS65987
     // === Power contracts ===
     struct Contract
     {
-        unsigned char role;
-        unsigned char USB_type;
-        unsigned int voltage;
-        unsigned int current;
+        uint8_t role;
+        uint8_t USB_type;
+        uint16_t voltage;
+        uint16_t current;
     };
 
     /**
@@ -90,7 +92,8 @@ namespace TPS65987
     class PDO
     {
     private:
-        unsigned long data{0}; // The raw data of the PDO
+        uint32_t data{0}; // The raw data of the PDO
+
     public:
         // ** Enum with PDO types ***
         enum Type
@@ -102,20 +105,128 @@ namespace TPS65987
         };
 
         // *** Constructors ***
-        PDO(void) = default;
-        PDO(PDO &other) = default;
+        PDO() = default;
+        explicit PDO(uint32_t data) : data(data) {}
+        PDO(const PDO &other) = default;
         PDO(PDO &&other) = default;
-        PDO &operator=(PDO &other) = default;
+        PDO &operator=(const PDO &other) = default;
         PDO &operator=(PDO &&other) = default;
-        PDO(unsigned long data) : data(data){};
+        PDO &operator=(uint32_t data)
+        {
+            this->data = data;
+            return *this;
+        }
+        ~PDO() = default;
 
         // *** Methods ***
-        unsigned long get_data(void) const;
-        unsigned int voltage(void) const;
-        unsigned int current(void) const;
-        Type  type(void) const;
-        void set_voltage(const unsigned int voltage);
-        void set_current(const unsigned int current);
+        [[nodiscard]] uint32_t get_data() const;
+        [[nodiscard]] uint16_t voltage() const;
+        [[nodiscard]] uint16_t current() const;
+        [[nodiscard]] Type type() const;
+        void set_voltage(uint16_t voltage);
+        void set_current(uint16_t current);
+    };
+
+    /**
+     * @brief Status bit field for non-interrupt events.
+     * @details Read-only register.
+     */
+    class Status : public reg_t
+    {
+    private:
+        std::array<uint8_t, Register::Status.length> data{0};
+    
+    public:
+        Status() : reg_t{Register::Status} {}
+
+        // *** Get all the options ***
+        [[nodiscard]] auto& get_data() { return this->data; }
+        [[nodiscard]] inline bool PlugPresent() const { return (this->data.at(0) & (1 << 0)) != 0; }
+        [[nodiscard]] inline uint8_t ConnState() const { return bits::get(this->data.at(0), {0b111, 1}); }
+        [[nodiscard]] inline bool PortRole() const { return (this->data.at(0) & (1 << 5)) != 0; }
+    };
+
+    /**
+     * @brief Configuration bits that define hardware that is common to all ports
+     *  and in most cases will not change in normal operation or will not
+     *  require immediate action if changed. Any modifications to this
+     *  register will cause a port disconnect and reconnect with the new
+     *  settings.
+     * @details Initialized by Application Customization. 
+     */
+    class GlobalConfiguration : public reg_t
+    {
+    private:
+        std::array<uint8_t, Register::Global_SysConfig.length> data{0};
+
+    public:
+        GlobalConfiguration() : reg_t{Register::Global_SysConfig} {}
+
+        // *** Get all the options ***
+        [[nodiscard]] auto& get_data() { return this->data; }
+        [[nodiscard]] inline uint8_t PP1_CABLEconfig() const { return bits::get(this->data.at(0), {0b11, 0}); }
+        [[nodiscard]] inline uint8_t PP2_CABLEconfig() const { return bits::get(this->data.at(0), {0b11, 2}); }
+        [[nodiscard]] inline uint8_t PP1config() const { return bits::get(this->data.at(2), {0b111, 0}); }
+        [[nodiscard]] inline uint8_t PP2config() const { return bits::get(this->data.at(2), {0b111, 3}); }
+
+        // *** Set all the options ***
+        inline void set_PP1_CABLEconfig(const uint8_t value) { this->data[0] = bits::set(this->data.at(0), {0b11, 0, value}); }
+        inline void set_PP2_CABLEconfig(const uint8_t value) { this->data[0] = bits::set(this->data.at(0), {0b11, 2, value}); }
+        inline void set_PP1config(const uint8_t value) { this->data[2] = bits::set(this->data.at(2), {0b111, 0, value}); }
+        inline void set_PP2config(const uint8_t value) { this->data[2] = bits::set(this->data.at(2), {0b111, 3, value}); }
+    };
+
+    /**
+     * @brief Configuration bits that define hardware that is specific to a the
+     *  respective port and in most cases will not change in normal
+     *  operation or will not require immediate action if changed. Any
+     *  modifications to this register will cause a port disconnect and
+     *  reconnect with the new settings.
+     * @details Initialized by Application Customization. 
+     */
+    class PortConfiguration : public reg_t
+    {
+    private:
+        std::array<uint8_t, Register::Port_Config.length> data{0};
+    
+    public:
+        PortConfiguration() : reg_t{Register::Port_Config} {}
+
+        // *** Get all the options ***
+        [[nodiscard]] auto& get_data() { return this->data; }
+        [[nodiscard]] inline uint8_t TypeCStateMachine() const { return bits::get(this->data.at(0), {0b11, 0}); }
+        [[nodiscard]] inline uint8_t ReceptacleType() const { return bits::get(this->data.at(0), {0b111, 3}); }
+        [[nodiscard]] inline uint8_t VCONNsupported() const { return bits::get(this->data.at(1), {0b11, 3}); }
+
+        // *** Set all the options ***
+        inline void set_TypeCStateMachine(const uint8_t value) { this->data[0] = bits::set(this->data.at(0), {0b11, 0, value}); }
+        inline void set_ReceptacleType(const uint8_t value) { this->data[0] = bits::set(this->data.at(0), {0b111, 3, value}); }
+        inline void set_VCONNsupported(const uint8_t value) { this->data[1] = bits::set(this->data.at(1), {0b11, 3, value}); }
+
+    };
+
+    /**
+     * @brief Configuration bits affecting system policy. These bits may change
+     *  during normal operation and are used for controlling the respective
+     *  port. The PD Controller will not take immediate action upon writing.
+     *  Changes made to this register will take effect the next time the
+     *  appropriate policy is invoked.
+     * @details Initialized by Application Customization. 
+     */
+    class PortControl : public reg_t
+    {
+    private:
+        std::array<uint8_t, Register::Port_Ctrl.length> data{0};
+
+    public:
+        PortControl() : reg_t{Register::Port_Ctrl} {}
+
+        // *** Get all the options ***
+        [[nodiscard]] auto& get_data() { return this->data; }
+        [[nodiscard]] inline uint8_t TypeCCurrent() const { return bits::get(this->data.at(0), {0b11, 0}); }
+
+        // *** Set all the options ***
+        inline void set_TypeCCurrent(const uint8_t value) { this->data[0] = bits::set(this->data.at(0), {0b11, 0, value}); }
     };
 
     using Bus::Data_t;
@@ -126,37 +237,84 @@ namespace TPS65987
         // *** properties ***
         bus_controller mybus;
         Data_t i2c_data = {0};
-        std::array<char, 6> buffer_cmd = {0x08, 0x04, 0, 0, 0, 0};
+        std::array<char, 6> buffer_cmd = {0x08, 0x04, 0, 0, 0, 0}; // NOLINT
         Mode mode_active = Mode::BOOT;
-        mutable char cmd_active[4] = {0};
+        mutable char cmd_active[4] = {0}; // NOLINT
         Contract contract_active = {0, 0, 0, 0};
 
     public:
-        std::array<unsigned char, 66> buffer_data = {0};
+        std::array<uint8_t, 66> buffer_data = {0}; // NOLINT
         // *** Constructor ***
         /**
          * @brief Constructor for PD controller.
          * @param bus_used The reference to the used bus peripheral.
          */
-        Controller(bus_controller bus_used)
-            : mybus{bus_used} {};
+        explicit Controller(bus_controller bus_used) : mybus{bus_used} {}
 
         // Properties
 
         // *** Methods ***
-        bool initialize(void);
-        Mode get_mode(void) const { return this->mode_active; };
-        char *get_active_command(void) const { return &this->cmd_active[0]; };
-        Contract get_active_contract(void) const { return this->contract_active; };
-        bool read_register(const reg_t reg);
-        bool write_register(const reg_t reg);
-        bool read_active_command(void);
+        bool initialize();
+        auto get_mode() const -> Mode { return this->mode_active; }
+        auto get_active_command() const -> char * { return &this->cmd_active[0]; }
+        auto get_active_contract() const -> Contract { return this->contract_active; }
+        bool read_register(reg_t reg);
+        bool write_register(reg_t reg);
+        bool read_active_command();
         bool write_command(const char *cmd);
-        bool read_mode(void);
-        bool read_PD_status(void);
-        std::optional<unsigned long> read_status(void);
-        std::optional<PDO> read_active_pdo(void);
-        std::optional<PDO> read_TX_sink_pdo(const unsigned char pdo_number);
+        bool read_mode();
+        bool read_PD_status();
+        auto read_status() -> std::optional<uint32_t>;
+        auto read_active_pdo() -> std::optional<PDO>;
+        auto read_TX_sink_pdo(uint8_t pdo_number) -> std::optional<PDO>;
+
+        /**
+         * @brief Read a register from the TPS65987 and directly
+         *        copy the data to the register class.
+         * 
+         * @tparam Treg The register class.
+         * @param reg The register object to copy the data to.
+         * @return Returns true if the register is read successfully.
+         */
+        template <class Treg>
+        bool read(Treg &reg)
+        {
+            // Read the register
+            if (!this->read_register({reg.address, reg.length}))
+                return false;
+
+            /* Copy the data to the register in reverse order
+             * because the i2c controller reads the data in reverse order.
+             */
+            std::copy(
+                this->buffer_data.rend() - reg.length - 1,
+                this->buffer_data.rend() - 1,
+                reg.get_data().begin()
+            );
+
+            return true;
+        }
+
+        /**
+         * @brief Write a register to the TPS65987.
+         * 
+         * @tparam Treg The register class.
+         * @param reg The register object to write.
+         * @return Returns true if the register is written successfully.
+         */
+        template <class Treg>
+        bool write(Treg &reg)
+        {
+            // Copy the data to the buffer
+            std::copy(
+                reg.get_data().cbegin(),
+                reg.get_data().cend(),
+                this->buffer_data.begin() + 2
+            );
+
+            // Write the register
+            return this->write_register({reg.address, reg.length});
+        }
     };
-};
+}; // namespace TPS65987
 #endif