+++ test/battery/test_tps65987/test_tps65987.cpp
@@ -31,48 +31,64 @@
 #include <unity.h>
 #include <mock.h>
 #include <array>
-#include "interface.h"
+#include <unordered_map>
+#include "interface.h" // NOLINT
 
 // === Fixtures ===
 
 // Mock the i2c driver
-struct I2C_Mock: public Driver::Base { };
+// NOLINTBEGIN
+struct I2C_Mock : public Driver::Base
+{
+};
 Mock::Callable<bool> set_target_address;
 Mock::Callable<bool> send_word;
 Mock::Callable<bool> send_array;
 Mock::Callable<bool> send_array_leader;
 Mock::Callable<bool> read_array;
-std::array<unsigned char, 66> rx_buffer{0};
+std::array<std::uint8_t, 66> rx_buffer{0};
 
-namespace Bus {
-    void change_address(I2C_Mock& bus, const unsigned char address)
+namespace Bus
+{
+    /**
+     * @attention The rx_buffer represents the written buffer verbatim!
+     * This means that the first byte of the buffer is the register address
+     * when sending a word or array. When reading an array, the first byte
+     * is the first received byte with the register address!
+     */
+    void change_address(I2C_Mock bus, const std::uint8_t address)
     {
         ::set_target_address(address);
         return;
-    };
-    bool send_word(I2C_Mock& bus, unsigned int word)
+    }
+    bool send_word(I2C_Mock bus, uint16_t word)
     {
-       return ::send_word(word);
-    };
-    bool send_array(I2C_Mock& bus, const unsigned char* data, const unsigned char n_bytes)
-    { 
+        return ::send_word(word);
+    }
+    bool send_array(I2C_Mock bus, const std::uint8_t *data, const std::uint8_t n_bytes)
+    {
         std::copy(data, data + n_bytes, rx_buffer.begin());
-        return ::send_array(n_bytes); 
-    };
-    bool send_array_leader(I2C_Mock& bus, const unsigned char byte, const unsigned char* data, const unsigned char n_bytes)
-    { 
-        return ::send_array_leader(byte, data, n_bytes); 
-    };
-    bool read_array(I2C_Mock &bus, const unsigned char reg, unsigned char* dest, const unsigned char n_bytes)
+        return ::send_array(n_bytes);
+    }
+    bool send_array_leader(
+        I2C_Mock bus,
+        const std::uint8_t byte,
+        const std::uint8_t *data,
+        const std::uint8_t n_bytes)
+    {
+        return ::send_array_leader(byte, data, n_bytes);
+    }
+    bool read_array(I2C_Mock bus, const std::uint8_t reg, std::uint8_t *dest, const std::uint8_t n_bytes)
     {
         std::copy(rx_buffer.begin(), rx_buffer.begin() + n_bytes, dest);
         return ::read_array(n_bytes);
-    };
-};
+    }
+}; // namespace Bus
+// NOLINTEND
 
 // Include the UUT
 #include "battery/tps65987.h"
-#include "battery/tps65987.cpp"
+#include "battery/tps65987.cpp" // NOLINT
 template class TPS65987::Controller<I2C_Mock>;
 
 /** === Test List ===
@@ -111,43 +127,45 @@ template class TPS65987::Controller<I2C_Mock>;
  *     ▢ 0x70 - Sleep Configuration
  */
 
-void setUp(void) {
+void setUp()
+{
     set_target_address.reset();
     send_word.reset();
     send_array.reset();
     send_array_leader.reset();
     read_array.reset();
-};
+}
 
-void tearDown(void) {
-// clean stuff up here
-};
+void tearDown()
+{
+    // clean stuff up here
+}
 
 // === Define Tests ===
 /// @brief Test the constructor
-void test_init(void)
+void test_init()
 {
     // Setup the mocked i2c driver
-    I2C_Mock i2c;
+    const I2C_Mock i2c;
 
     // create the controller object
-    TPS65987::Controller UUT(i2c);
+    const TPS65987::Controller UUT(i2c);
 
     // perform testing
-    TEST_ASSERT_EQUAL(0, static_cast<unsigned char>(UUT.get_mode()));
-    char Expected_Command[] = {0,0,0,0};
+    TEST_ASSERT_EQUAL(0, static_cast<std::uint8_t>(UUT.get_mode()));
+    char Expected_Command[] = {0, 0, 0, 0}; // NOLINT
     TEST_ASSERT_EQUAL_CHAR_ARRAY(Expected_Command, UUT.get_active_command(), 4);
     TEST_ASSERT_EQUAL(0, UUT.get_active_contract().role);
     TEST_ASSERT_EQUAL(0, UUT.get_active_contract().USB_type);
     TEST_ASSERT_EQUAL(0, UUT.get_active_contract().voltage);
     TEST_ASSERT_EQUAL(0, UUT.get_active_contract().current);
-};
+}
 
 /// @brief Test reading a register with variable length
-void test_read_register(void)
+void test_read_register()
 {
     // Setup the mocked i2c driver
-    I2C_Mock i2c;
+    const I2C_Mock i2c;
 
     // create the controller object
     TPS65987::Controller UUT(i2c);
@@ -158,13 +176,13 @@ void test_read_register(void)
 
     TEST_ASSERT_TRUE(UUT.read_register(TPS65987::Register::Cmd1));
     ::read_array.assert_called_once_with(5);
-};
+}
 
 /// @brief Test writing a register with variable length
-void test_write_register(void)
+void test_write_register()
 {
     // Setup the mocked i2c driver
-    I2C_Mock i2c;
+    const I2C_Mock i2c;
 
     // create the controller object
     TPS65987::Controller UUT(i2c);
@@ -175,13 +193,13 @@ void test_write_register(void)
 
     TEST_ASSERT_TRUE(UUT.write_register(TPS65987::Register::Cmd1));
     ::send_array.assert_called_once_with(6);
-};
+}
 
 /// @brief Test reading a command status
-void test_read_active_command(void)
+void test_read_active_command()
 {
     // Setup the mocked i2c driver
-    I2C_Mock i2c;
+    const I2C_Mock i2c;
 
     // create the controller object
     TPS65987::Controller UUT(i2c);
@@ -194,13 +212,13 @@ void test_read_active_command(void)
     TEST_ASSERT_TRUE(UUT.read_active_command());
     ::read_array.assert_called_once_with(5);
     TEST_ASSERT_EQUAL_CHAR_ARRAY("CMD!", UUT.get_active_command(), 4);
-};
+}
 
 /// @brief Test writting a command
-void test_write_command(void)
+void test_write_command()
 {
     // Setup the mocked i2c driver
-    I2C_Mock i2c;
+    const I2C_Mock i2c;
 
     // create the controller object
     TPS65987::Controller UUT(i2c);
@@ -208,13 +226,13 @@ void test_write_command(void)
     // perform testing
     TEST_ASSERT_TRUE(UUT.write_command("PTCc"));
     TEST_ASSERT_EQUAL_CHAR_ARRAY("PTCc", &rx_buffer[2], 4);
-};
+}
 
 /// @brief Test reading the current mode of the controller
-void test_read_mode(void)
+void test_read_mode()
 {
     // Setup the mocked i2c driver
-    I2C_Mock i2c;
+    const I2C_Mock i2c;
 
     // create the controller object
     TPS65987::Controller UUT(i2c);
@@ -226,14 +244,14 @@ void test_read_mode(void)
     rx_buffer[3] = 'P';
     TEST_ASSERT_TRUE(UUT.read_mode());
     ::read_array.assert_called_once_with(5);
-    TEST_ASSERT_EQUAL(1, static_cast<unsigned char>(UUT.get_mode()));
-};
+    TEST_ASSERT_EQUAL(1, static_cast<std::uint8_t>(UUT.get_mode()));
+}
 
 /// @brief Test the correct initialization of the controller
-void test_initialization(void)
+void test_initialization()
 {
     // Setup the mocked i2c driver
-    I2C_Mock i2c;
+    const I2C_Mock i2c;
 
     // create the controller object
     TPS65987::Controller UUT(i2c);
@@ -248,13 +266,13 @@ void test_initialization(void)
     ::set_target_address.assert_called_once_with(TPS65987::i2c_address);
     ::read_array.assert_called_once_with(5);
     TEST_ASSERT_EQUAL_CHAR_ARRAY("PTCc", &rx_buffer[2], 4);
-};
+}
 
 /// @brief Test the reading of the PD status
-void test_read_PD_status(void)
+void test_read_PD_status()
 {
     // Setup the mocked i2c driver
-    I2C_Mock i2c;
+    const I2C_Mock i2c;
 
     // create the controller object
     TPS65987::Controller UUT(i2c);
@@ -325,13 +343,13 @@ void test_read_PD_status(void)
     TEST_ASSERT_EQUAL(3, UUT.get_active_contract().USB_type);
     TEST_ASSERT_EQUAL(0, UUT.get_active_contract().voltage);
     TEST_ASSERT_EQUAL(0, UUT.get_active_contract().current);
-};
+}
 
 /// @brief Test the reading of the status
-void test_read_status(void)
+void test_read_status()
 {
     // Setup the mocked i2c driver
-    I2C_Mock i2c;
+    const I2C_Mock i2c;
 
     // create the controller object
     TPS65987::Controller UUT(i2c);
@@ -348,11 +366,11 @@ void test_read_status(void)
     auto response = UUT.read_status();
     TEST_ASSERT_TRUE(response);
     TEST_ASSERT_EQUAL(0x44332211, response.value());
-    ::read_array.assert_called_once_with(8+1);
-};
+    ::read_array.assert_called_once_with(8 + 1);
+}
 
 /// @brief Test the PDO class for USB PD contracts
-void test_PDO_class(void)
+void test_PDO_class()
 {
     // Create a PDO object
     TPS65987::PDO pdo;
@@ -361,7 +379,7 @@ void test_PDO_class(void)
     TEST_ASSERT_EQUAL(0, pdo.voltage());
     TEST_ASSERT_EQUAL(0, pdo.current());
     TEST_ASSERT_EQUAL(0, pdo.get_data());
-    TEST_ASSERT_EQUAL( TPS65987::PDO::Fixed_Supply, pdo.type() );
+    TEST_ASSERT_EQUAL(TPS65987::PDO::Fixed_Supply, pdo.type());
 
     // Test setting the voltage
     pdo.set_voltage(5000);
@@ -374,31 +392,31 @@ void test_PDO_class(void)
     TEST_ASSERT_EQUAL(3000, pdo.current());
 
     // Test constructor with data
-    TPS65987::PDO pdo2{(0b11 << 30) | 0x1912CUL};
-    TEST_ASSERT_EQUAL( TPS65987::PDO::APDO, pdo2.type() );
+    TPS65987::PDO pdo2{(0b11UL << 30) | 0x1912CUL};
+    TEST_ASSERT_EQUAL(TPS65987::PDO::APDO, pdo2.type());
 
     // Test copy constructor
-    TPS65987::PDO pdo3{pdo};
+    const TPS65987::PDO pdo3{pdo};
     TEST_ASSERT_EQUAL(5000, pdo3.voltage());
     TEST_ASSERT_EQUAL(3000, pdo3.current());
 
     // Test assignment operator
     TPS65987::PDO pdo4;
-    pdo4 = 0x1912C;
+    pdo4 = 0x1912CUL;
     TEST_ASSERT_EQUAL(5000, pdo4.voltage());
     TEST_ASSERT_EQUAL(3000, pdo4.current());
 
     // Test getting the voltage of a variable supply
     pdo2 = (0b10UL << 30) | (0x190UL << 20);
-    TEST_ASSERT_EQUAL( TPS65987::PDO::Variable_Supply, pdo2.type() );
-    TEST_ASSERT_EQUAL( 20000, pdo2.voltage() );
-};
+    TEST_ASSERT_EQUAL(TPS65987::PDO::Variable_Supply, pdo2.type());
+    TEST_ASSERT_EQUAL(20000, pdo2.voltage());
+}
 
 /// @brief Test the reading of the active contract PDO
-void test_read_active_pdo(void)
+void test_read_active_pdo()
 {
     // Setup the mocked i2c driver
-    I2C_Mock i2c;
+    const I2C_Mock i2c;
 
     // create the controller object
     TPS65987::Controller UUT(i2c);
@@ -413,14 +431,14 @@ void test_read_active_pdo(void)
     auto response = UUT.read_active_pdo();
     TEST_ASSERT_TRUE(response);
     TEST_ASSERT_EQUAL_HEX(0x44332211, response.value().get_data());
-    ::read_array.assert_called_once_with(6+1);
-};
+    ::read_array.assert_called_once_with(6 + 1);
+}
 
-/// @brief Test the reading of the TX sink capabilites PDOs
-void test_read_TX_sink_capability(void)
+/// @brief Test the reading of the TX sink capabilities PDOs
+void test_read_TX_sink_capability()
 {
     // Setup the mocked i2c driver
-    I2C_Mock i2c;
+    const I2C_Mock i2c;
 
     // create the controller object
     TPS65987::Controller UUT(i2c);
@@ -434,11 +452,172 @@ void test_read_TX_sink_capability(void)
     auto response = UUT.read_TX_sink_pdo(0);
     TEST_ASSERT_TRUE(response);
     TEST_ASSERT_EQUAL_HEX(0x55443322, response.value().get_data());
-    ::read_array.assert_called_once_with(57+1);
-};
+    ::read_array.assert_called_once_with(57 + 1);
+}
+
+/// @brief Test the status register class
+void test_class_status()
+{
+    // Create empty register
+    TPS65987::Status status;
+
+    // Test getting the underlying data array
+    auto &data = status.get_data();
+    TEST_ASSERT_EQUAL(TPS65987::Register::Status.length, data.size());
+
+    // Test reading the register options
+    TEST_ASSERT_EQUAL(TPS65987::Register::Status.address, status.address);
+    TEST_ASSERT_EQUAL(TPS65987::Register::Status.length, status.length);
+
+    // Test reading the plug present bit
+    TEST_ASSERT_FALSE(status.PlugPresent());
+    data[0] = 0x01;
+    TEST_ASSERT_TRUE(status.PlugPresent());
+
+    // Test reading the ConnState property
+    TEST_ASSERT_EQUAL(0, status.ConnState());
+    data[0] |= 0x02 << 1;
+    TEST_ASSERT_EQUAL(2, status.ConnState());
+
+    // Test reading the port role
+    TEST_ASSERT_FALSE(status.PortRole());
+    data[0] |= 0x01 << 5;
+    TEST_ASSERT_TRUE(status.PortRole());
+}
+
+/// @brief Test the specialized register classes
+void test_class_global_system_configuration()
+{
+    // Create empty register
+    TPS65987::GlobalConfiguration global_config;
+
+    // Test getting the underlying data array
+    auto data = global_config.get_data();
+    TEST_ASSERT_EQUAL(TPS65987::Register::Global_SysConfig.length, data.size());
+
+    // Test reading the register options
+    TEST_ASSERT_EQUAL(TPS65987::Register::Global_SysConfig.address, global_config.address);
+    TEST_ASSERT_EQUAL(TPS65987::Register::Global_SysConfig.length, global_config.length);
+    TEST_ASSERT_EQUAL(0, global_config.PP1_CABLEconfig());
+    TEST_ASSERT_EQUAL(0, global_config.PP2_CABLEconfig());
+    TEST_ASSERT_EQUAL(0, global_config.PP1config());
+    TEST_ASSERT_EQUAL(0, global_config.PP2config());
+
+    // Test setting the register options
+    global_config.set_PP1_CABLEconfig(1);
+    global_config.set_PP2_CABLEconfig(1);
+    global_config.set_PP1config(1);
+    global_config.set_PP2config(1);
+    TEST_ASSERT_EQUAL(1, global_config.PP1_CABLEconfig());
+    TEST_ASSERT_EQUAL(1, global_config.PP2_CABLEconfig());
+    TEST_ASSERT_EQUAL(1, global_config.PP1config());
+    TEST_ASSERT_EQUAL(1, global_config.PP2config());
+    TEST_ASSERT_EQUAL(0b1001, global_config.get_data().at(2));
+
+    // Test copying an array into the register
+    std::array<uint8_t, 14> raw{0};
+    raw.fill(0x01);
+    std::copy(raw.begin(), raw.end(), global_config.get_data().begin());
+    TEST_ASSERT_EQUAL(0x01, global_config.get_data().at(0));
+    TEST_ASSERT_EQUAL(0x01, global_config.get_data().at(13));
+}
+
+/// @brief Test reading the Global System Configuration Register
+void test_reading_global_system_configuration()
+{
+    // Setup the mocked i2c driver
+    const I2C_Mock i2c;
+
+    // create the controller object
+    TPS65987::Controller UUT(i2c);
+
+    // Set up response
+    rx_buffer[11] = 0x01; // Byte 2
+    rx_buffer[12] = 0x00; // Byte 1
+    rx_buffer[13] = 0x00; // Byte 0
+    rx_buffer[14] = 14; // Length
+
+    // Read the register
+    TPS65987::GlobalConfiguration global_config;
+    auto response = UUT.read(global_config);
+
+    // Test the register options
+    TEST_ASSERT_TRUE(response);
+    TEST_ASSERT_EQUAL(1, global_config.PP1config());
+    TEST_ASSERT_EQUAL(0, global_config.PP2config());
+}
+
+/// @brief Test writing the Global System Configuration Register
+void test_writing_global_system_configuration()
+{
+    // Setup the mocked i2c driver
+    const I2C_Mock i2c;
+
+    // create the controller object
+    TPS65987::Controller UUT(i2c);
+
+    // Set and write the register content
+    TPS65987::GlobalConfiguration global_config;
+    global_config.set_PP2config(3);
+    auto response = UUT.write(global_config);
+
+    // Test the content of the send data
+    TEST_ASSERT_TRUE(response);
+    TEST_ASSERT_EQUAL(0x27, rx_buffer.at(0));
+    TEST_ASSERT_EQUAL(14, rx_buffer.at(1));
+    TEST_ASSERT_EQUAL(0, rx_buffer.at(2));
+    TEST_ASSERT_EQUAL(0, rx_buffer.at(3));
+    TEST_ASSERT_EQUAL((3<<3), rx_buffer.at(4));
+}
+
+/// @brief Test the port configuration register class
+void test_class_port_configuration()
+{
+    // Create empty register
+    TPS65987::PortConfiguration port_config;
+
+    // Test getting the underlying data array
+    auto data = port_config.get_data();
+    TEST_ASSERT_EQUAL(TPS65987::Register::Port_Config.length, data.size());
+
+    // Test reading the register options
+    TEST_ASSERT_EQUAL(TPS65987::Register::Port_Config.address, port_config.address);
+    TEST_ASSERT_EQUAL(TPS65987::Register::Port_Config.length, port_config.length);
+    TEST_ASSERT_EQUAL(0, port_config.TypeCStateMachine());
+    TEST_ASSERT_EQUAL(0, port_config.ReceptacleType());
+    TEST_ASSERT_EQUAL(0, port_config.VCONNsupported());
+
+    // Test setting the register options
+    port_config.set_TypeCStateMachine(1);
+    port_config.set_ReceptacleType(1);
+    port_config.set_VCONNsupported(1);
+    TEST_ASSERT_EQUAL(1, port_config.TypeCStateMachine());
+    TEST_ASSERT_EQUAL(1, port_config.ReceptacleType());
+    TEST_ASSERT_EQUAL(1, port_config.VCONNsupported());
+}
+
+/// @brief Test the port control register class
+void test_class_port_control()
+{
+    // Create empty register
+    TPS65987::PortControl port_control;
+
+    // Test getting the underlying data array
+    auto data = port_control.get_data();
+    TEST_ASSERT_EQUAL(TPS65987::Register::Port_Ctrl.length, data.size());
+    TEST_ASSERT_EQUAL(0, port_control.TypeCCurrent());
+
+    // Test reading the register options
+    TEST_ASSERT_EQUAL(TPS65987::Register::Port_Ctrl.address, port_control.address);
+    TEST_ASSERT_EQUAL(TPS65987::Register::Port_Ctrl.length, port_control.length);
+
+    // Test setting the register options
+    port_control.set_TypeCCurrent(1);
+    TEST_ASSERT_EQUAL(1, port_control.TypeCCurrent());
+}
 
 /// === Run Tests ===
-int main(int argc, char** argv)
+int main()
 {
     UNITY_BEGIN();
     RUN_TEST(test_init);
@@ -454,5 +633,11 @@ int main(int argc, char** argv)
     RUN_TEST(test_PDO_class);
     RUN_TEST(test_read_active_pdo);
     RUN_TEST(test_read_TX_sink_capability);
+    RUN_TEST(test_class_status);
+    RUN_TEST(test_class_global_system_configuration);
+    RUN_TEST(test_reading_global_system_configuration);
+    RUN_TEST(test_writing_global_system_configuration);
+    RUN_TEST(test_class_port_configuration);
+    RUN_TEST(test_class_port_control);
     return UNITY_END();
-};
+}